faster database queries and loggin and more

+++ Http/controllers/students/applications/destroy.php
@@ -14,6 +14,13 @@ if ($application['interview_id'] !== null) {
     redirect('/students/applications');
 }
 
+//dd($application['is_second_round']);
+
+if ($application['is_second_round'] === true) {
+    Session::flash('toast', 'You cannot delete this application because it is a second round application');
+    redirect('/students/applications');
+}
+
 Application::delete($id);
 
 redirect('/students/applications');
